accept, decline

+++ controllers/onhold.js
@@ -25,38 +25,49 @@ exports.Join = asyncHandler(async (req, res, next) => {
 exports.Acceptjoin = asyncHandler(async (req, res, next) => {
   const id = req.params.id;
   const onHold = await Onhold.findById(id);
+  const owner = req.user
   if (!onHold) {
     return next(new ErrorResponse(`Join is not find with id ${id}`));
   }
-  const event = await Event.findByIdAndUpdate(onHold.event, {
-    $pull: { request: id },
-  });
-  const added = await Event.findByIdAndUpdate(onHold.event, {
-    $push: { members: onHold.user },
-  });
-  const removed = await Onhold.findByIdAndDelete(id);
-  res.status(200).json({
-    success: true,
-    data: `User accepted for event`,
-  });
+  const event = await Event.findById(onHold.event)
+  if(event.createdBy === owner){
+    const event = await Event.findByIdAndUpdate(onHold.event, {
+      $pull: { request: id },
+    });
+    const added = await Event.findByIdAndUpdate(onHold.event, {
+      $push: { members: onHold.user },
+    });
+    const removed = await Onhold.findByIdAndDelete(id);
+    res.status(200).json({
+      success: true,
+      data: `User accepted for event`,
+    });
+  } else {
+    return next(new ErrorResponse(`Not your Event`));
+  }
 });
 
 // decline join
 exports.Declinejoin = asyncHandler(async (req, res, next) => {
   const id = req.params.id;
   const onHold = await Onhold.findById(id);
+  const owner = req.user
   if (!onHold) {
     return next(new ErrorResponse(`Join is not find with id ${id}`));
   }
-  const event = await Event.findByIdAndUpdate(onHold.event, {
-    $pull: { request: id },
-  });
-  const added = await Event.findByIdAndUpdate(onHold.event, {
-    $push: { declined: onHold.user },
-  });
-  const removed = await Onhold.findByIdAndDelete(id);
-  res.status(200).json({
-    success: true,
-    data: `User declined for event`,
-  });
+  const event = await Event.findById(onHold.event)
+  if(event.createdBy === owner){
+    const event = await Event.findByIdAndUpdate(onHold.event, {
+      $pull: { request: id },
+    });
+    const added = await Event.findByIdAndUpdate(onHold.event, {
+      $push: { declined: onHold.user },
+    });
+    const removed = await Onhold.findByIdAndDelete(id);
+    res.status(200).json({
+      success: true,
+      data: `User declined for event`,
+    });
+  }
+  return next(new ErrorResponse(`Not your Event`));
 });
